pyproject.toml
[project]
name = "expenses_app"
version = "0.0.1"
authors = [
    { name="Matheus Vitorino", email="matheusvs122@gmail.com" },
]
description = "Small app to track and analyze personal expenses"
requires-python = ">=3.7"
dependencies = [
    "flask",
    "flask-cors",
    "flask-sqlalchemy",
    "sqlalchemy",
    "sqlalchemy-orm",
    "marshmallow_sqlalchemy",
    "cryptography",
    "ariadne",
    "pyjwt"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"